decimal replaced by double type

+++ Metrics-Analyzer/Data/CSV/CSV_AppMetrics.cs
@@ -5,8 +5,8 @@ internal class CSV_AppMetrics
     public DateTime date            { private set; get; }
     public string   app_name        { private set; get; }
     public int      company_id      { private set; get; }
-    public decimal  revenue         { private set; get; }
-    public decimal  marketing_spend { private set; get; }
+    public double   revenue         { private set; get; }
+    public double   marketing_spend { private set; get; }
 
     public static List<CSV_AppMetrics> Parse(string filePath)
     {
@@ -22,11 +22,11 @@ internal class CSV_AppMetrics
                 var columns = line.Split(CSVUtils.ColumnSeparator);
                 return new CSV_AppMetrics
                 {
-                    date            = CSVUtils.ParseDate    (columns[0]),
-                    app_name        = CSVUtils.ParseString  (columns[1]),
-                    company_id      = CSVUtils.ParseInt     (columns[2]),
-                    revenue         = CSVUtils.ParseDecimal (columns[3]),
-                    marketing_spend = CSVUtils.ParseDecimal (columns[4])
+                    date            = CSVUtils.ParseDate   (columns[0]),
+                    app_name        = CSVUtils.ParseString (columns[1]),
+                    company_id      = CSVUtils.ParseInt    (columns[2]),
+                    revenue         = CSVUtils.ParseDouble (columns[3]),
+                    marketing_spend = CSVUtils.ParseDouble (columns[4])
                 };
             })
             .ToList();
